requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
filelock==3.29.0
fsspec==2026.4.0
idna==3.13
Jinja2==3.1.6
kagglehub==1.0.1
kagglesdk==0.1.23
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pydicom==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
